scripts_com_sql/requirements.txt
bcrypt==4.2.0
cffi==1.17.0
cryptography==43.0.0
et-xmlfile==1.1.0
greenlet==3.0.3
kaleido==0.2.1
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pillow==10.4.0
plotly==5.24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.32
tenacity==9.0.0
typing-extensions==4.12.2
tzdata==2024.1
